requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LangChain / LangGraph
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-core==0.2.22
langchain-text-splitters==0.2.1
langgraph==0.1.19

# Vector DB
chromadb==0.5.3

# Document loaders
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.4

# Web scraping
beautifulsoup4==4.12.3
lxml==5.2.2

# Structured DB
sqlalchemy==2.0.30

# Guardrails / NLP
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
# spacy model installed separately: python -m spacy download en_core_web_sm

# Utilities
pydantic==2.7.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.1.0
tiktoken==0.7.0
numpy==1.26.4